Rearrange consumption module
